fix fit all columns to content block error when upgrading
