+++ src/block-components/content-align/attributes.js
@@ -1,4 +1,8 @@
+import { deprecatedAddAttributes } from './deprecated'
+
 export const addAttributes = attrObject => {
+	deprecatedAddAttributes( attrObject )
+
 	attrObject.add( {
 		attributes: {
 			innerBlockContentAlign: {
