+++ src/block/columns/deprecated.js
@@ -34,14 +34,14 @@ addFilter( 'stackable.columns.save.contentClassNames', 'stackable/3.6.2', ( clas
 const deprecated = [
 	{
 		attributes: attributes(),
-		save: withVersion( '3.8.0' )( Save ),
+		save: withVersion( '3.7.3' )( Save ),
 		isEligible: attributes => !! attributes.columnFit,
 		migrate: attributes => {
 			return {
 				...attributes,
 				columnFit: '',
 				columnFitAlign: '',
-				columnJustify: attributes.columnFitAlign,
+				columnJustify: !! attributes.columnFit ? ( attributes.columnFitAlign || 'flex-start' ) : '',
 			}
 		},
 	},
